Changed protocol dclass option to use only id instead of classes for properties
This is probably the best way to go since this option is largely used for reverse engineering classes. The class could reference other classes as properties that are stripped. Until I write a recursive protocol generator for the classes that are referenced as properties, I’ll just cast them as id’s for now