Adds Data Directories Paths script
The `ddp` script provides access to obtaining the common paths for an
application bundle. The directories that can be obtained are:

- DocumentDirectory
- Shared directories from application groups

The script also provides a method to display the application groups the
app belongs to. DPP makes use of the `SecTaskCopyValueForEntitlement`
function. As explained in [Stack Overflow Answer](https://stackoverflow.com/a/53378031/8669826).